refactoring: add interfaces, rename bad names, remove extra logic from view and minor fixes
