+++ src/interfaces/repository.py
@@ -1,0 +1,11 @@
+from abc import ABC, abstractmethod
+
+
+class IRepository(ABC):
+    @abstractmethod
+    def load(self) -> dict:
+        raise NotImplementedError
+
+    @abstractmethod
+    def save(self, data: dict):
+        raise NotImplementedError
