--- src/model/validation.py
@@ -1,31 +1,0 @@
-import re
-from pathlib import Path
-
-
-def is_phone_number(number: str) -> bool:
-    """
-    Проверка строки на принадлежность к телефонному номеру.
-    Телефонный номер должен иметь спец символ + в начале,
-    должен быть длинной от 7 до 15 символов, включая код страны
-    и может иметь спец символы (Скобки, пробельные символы, дефисы и тире)
-    :param number: Проверяемая строка с телефонным номером
-    :return: True, если строка является телефонным номером, иначе False
-    """
-    # Список разрешенных спец-символов в телефонном номере
-    # Скобки, пробельные символы, дефисы и тире
-    pattern = r'[\(\)\s\u002D\u2010\u2011\u2012\u2013\u2014\u2212]+'
-    # Отсутствует номер
-    if not number:
-        return False
-    # Пропущен указатель на международный код страны
-    if len(number) > 0 and number[0] != '+':
-        return False
-    number = number[1:]
-    number = re.sub(pattern, '', number)
-    # В номере присутствуют недопустимые символы
-    if not number.isdigit():
-        return False
-    # Некорректная длина телефонного номера
-    if not (7 < len(number) < 15):
-        return False
-    return True
